# server/qualiva-openapi.yaml
openapi: 3.0.3
info:
  title: Qualiva - Enterprise Secure Module Management System (ESMMS) API
  description: |
    Comprehensive OpenAPI 3.x specification for the Qualiva ESMMS backend services.
    Generated strictly via comprehensive source code inspection of `server/src` for API security auditing with OWASP ZAP.
    
    Includes role-based access control (RBAC), multi-engine database proxies, granular moderator permissions,
    proctored assessment workflows, and real-time enterprise communication endpoints.
  version: 1.0.0
servers:
  - url: http://host.docker.internal:5050
    description: Local development backend server
  - url: /
    description: Current host relative path
...
paths:
...
  /api/v1/chat/upload:
    post:
      tags:
        - Real-Time Chat & Inbox
      summary: Upload Chat File Attachment
      description: Multer-based file upload up to 50MB for chat attachments.
      operationId: uploadChatFile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: File uploaded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  file:
                    type: object
                    properties:
                      fileUrl:
                        type: string
                      fileName:
                        type: string
                      fileType:
                        type: string
                      fileSize:
                        type: integer
        '400':
          description: No file uploaded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/chat/community/threads:
    get:
      tags:
        - Real-Time Chat & Inbox
      summary: List Support Desk User Threads (Staff Only)
      description: Returns aggregated community support threads for Administrators and Moderators.
      operationId: getCommunityThreads
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Support threads.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  threads:
                    type: array
                    items:
                      type: object
        '403':
          description: Forbidden (Staff access required).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/chat/read-announcements:
    post:
      tags:
        - Real-Time Chat & Inbox
      summary: Mark Global Announcements as Read
      description: Updates the caller's last announcement read timestamp.
      operationId: markAnnouncementsRead
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Announcements marked read.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardSuccessResponse'
...
  /api/v1/notifications:
    get:
      tags:
        - Notifications
      summary: List Persistent Notifications
      description: Returns in-app notifications targeted to the caller's user ID or RBAC role.
      operationId: getNotifications
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            maximum: 50
      responses:
        '200':
          description: Notifications list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  unreadCount:
                    type: integer
                  notifications:
                    type: array
                    items:
                      $ref: '#/components/schemas/Notification'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JSON Web Token passed in the `Authorization` header as `Bearer <token>`.
        Payload contains `sub` (User UUID), `email`, `role`, `status`, and `department`.
    cookieAuth:
      type: apiKey
      in: cookie
      name: esmms_refresh_token
      description: |
        HttpOnly cookie containing the cryptographic refresh token used exclusively by `/api/v1/auth/refresh` and `/api/v1/auth/logout`.
  schemas:
    StandardSuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully."
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "Invalid request payload."
        details:
          type: array
          items:
            type: string
...
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        recipientId:
          type: string
          nullable: true
        targetRole:
          type: string
          nullable: true
        title:
          type: string
        message:
          type: string
        type:
          type: string
        link:
          type: string
          nullable: true
        isRead:
          type: boolean
        metadata:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...