# server/src/controllers/chatController.js
import { prisma } from '../utils/prisma.js';
import { logger } from '../utils/logger.js';
import { io } from '../server.js';
import xss from 'xss';

const userLastViewedAnnouncements = new Map();

async function getLastAnnouncementRead(userId) {
  try {
    const mem = userLastViewedAnnouncements.get(userId);
    if (mem) return mem;
    const record = await prisma.notification.findFirst({
      where: { recipientId: userId, type: 'LAST_ANNOUNCEMENT_READ' },
      orderBy: { updatedAt: 'desc' },
    });
    if (record) {
      const dt = new Date(record.updatedAt);
      userLastViewedAnnouncements.set(userId, dt);
      return dt;
    }
    return null;
  } catch {
    return null;
  }
}

async function setLastAnnouncementRead(userId) {
  try {
    const now = new Date();
    userLastViewedAnnouncements.set(userId, now);
    const existing = await prisma.notification.findFirst({
      where: { recipientId: userId, type: 'LAST_ANNOUNCEMENT_READ' },
    });
    if (existing) {
      await prisma.notification.update({
        where: { id: existing.id },
        data: { updatedAt: now, isRead: true },
      });
    } else {
      await prisma.notification.create({
        data: {
          recipientId: userId,
          type: 'LAST_ANNOUNCEMENT_READ',
          title: 'Last Seen Announcement',
          message: 'User read announcements',
          isRead: true,
        },
      });
    }
  } catch (err) {
    logger.warn(`Failed to persist last announcement read: ${err.message}`);
  }
}

/**
 * GET /api/v1/chat/messages
 * Fetch messages for ANNOUNCEMENT, COMMUNITY, or DIRECT channels
 */
export async function getMessages(req, res) {
  try {
    const { type = 'ANNOUNCEMENT', contactId, audit = 'false' } = req.query;
    const userId = req.user.id;
    const userRole = req.user.role;

    let where = { type };

    if (type === 'COMMUNITY') {
      // Community Support is a private support desk between this user and Admin/Moderators
      if (userRole === 'ADMIN' || userRole === 'MODERATOR') {
        if (contactId) {
          // Admin/Mod viewing or replying to a specific user's support thread
          where = {
            type: 'COMMUNITY',
            OR: [
              { senderId: contactId },
              { recipientId: contactId },
            ],
          };
        } else {
          // Admin/Mod viewing general unassigned support queue
          where = { type: 'COMMUNITY' };
        }
      } else {
        // Students/Creators only see their own tickets
        where = {
          type: 'COMMUNITY',
          OR: [
            { senderId: userId },
            { recipientId: userId },
          ],
        };
      }
    } else if (type === 'DIRECT') {
      if (audit === 'true' && (userRole === 'ADMIN' || userRole === 'MODERATOR')) {
        where = { type: 'DIRECT' };
      } else {
        if (!contactId) {
          return res.status(200).json({ success: true, count: 0, messages: [] });
        }
        where = {
          type: 'DIRECT',
          OR: [
            { senderId: userId, recipientId: contactId },
            { senderId: contactId, recipientId: userId },
          ],
        };
      }
    }

    const messages = await prisma.chatMessage.findMany({
      where,
      orderBy: { createdAt: 'asc' },
      take: 100,
    });

    // Mark unread direct messages as read
    if (type === 'DIRECT' && contactId) {
      await prisma.chatMessage.updateMany({
        where: {
          type: 'DIRECT',
          senderId: contactId,
          recipientId: userId,
          read: false,
        },
        data: { read: true },
      });
    } else if (type === 'COMMUNITY') {
      if (userRole === 'ADMIN' || userRole === 'MODERATOR') {
        if (contactId) {
          await prisma.chatMessage.updateMany({
            where: {
              type: 'COMMUNITY',
              senderId: contactId,
              read: false,
            },
            data: { read: true },
          });
        }
      } else {
        await prisma.chatMessage.updateMany({
          where: {
            type: 'COMMUNITY',
            senderRole: { in: ['ADMIN', 'MODERATOR'] },
            OR: [
              { recipientId: userId },
              { recipientId: null },
            ],
            read: false,
          },
          data: { read: true },
        });
      }
    } else if (type === 'ANNOUNCEMENT') {
      await setLastAnnouncementRead(userId);
    }

    return res.status(200).json({
      success: true,
      count: messages.length,
      messages,
    });
  } catch (error) {
    logger.error(`Get Chat Messages Error: ${error.message}`);
    return res.status(500).json({ success: false, error: 'FETCH_FAILED', message: 'Failed to retrieve chat messages.' });
  }
}

/**
 * POST /api/v1/chat/upload
 * Upload document or media attachment for chat
 */
export async function uploadChatFile(req, res) {
  try {
    if (!req.file) {
      return res.status(400).json({ success: false, error: 'NO_FILE', message: 'No file uploaded.' });
    }
    const fileUrl = `/uploads/chat/${req.file.filename}`;
    return res.status(200).json({
      success: true,
      file: {
        fileUrl,
        fileName: req.file.originalname,
        fileType: req.file.mimetype,
        fileSize: req.file.size,
      },
    });
  } catch (error) {
    logger.error(`Chat File Upload Error: ${error.message}`);
    return res.status(500).json({ success: false, error: 'UPLOAD_FAILED', message: 'Failed to upload chat file.' });
  }
}

/**
 * POST /api/v1/chat/messages
 * Send message to ANNOUNCEMENT (Admin only), COMMUNITY, or DIRECT
 */
export async function postMessage(req, res) {
  try {
    const { type, content, recipientId, fileUrl, fileName, fileType, fileSize, replyTo } = req.body;
    const user = req.user;

    const trimmedContent = content?.trim() || '';
    if (!trimmedContent && !fileUrl) {
      return res.status(400).json({
        success: false,
        error: 'EMPTY_CONTENT',
        message: 'Message content or file attachment is required.',
      });
    }

    if (type === 'ANNOUNCEMENT' && user.role !== 'ADMIN') {
      return res.status(403).json({
        success: false,
        error: 'FORBIDDEN',
        message: 'Only system administrators can broadcast global announcements.',
      });
    }

    let targetUser = null;
    if (recipientId) {
      targetUser = await prisma.user.findUnique({
        where: { id: recipientId },
        select: { id: true, name: true, role: true },
      });
      if (!targetUser && type === 'DIRECT') {
        return res.status(404).json({ success: false, error: 'USER_NOT_FOUND', message: 'Recipient not found.' });
      }
    } else if (type === 'DIRECT') {
      return res.status(400).json({ success: false, error: 'MISSING_RECIPIENT', message: 'Recipient is required for 1-on-1 direct messaging.' });
    }

    const sanitizedContent = trimmedContent ? xss(trimmedContent) : (fileName || 'Attachment');

    const message = await prisma.chatMessage.create({
      data: {
        type: type || 'COMMUNITY',
        senderId: user.id,
        senderName: user.name,
        senderEmail: user.email,
        senderRole: user.role,
        recipientId: targetUser?.id || null,
        recipientName: targetUser?.name || null,
        recipientRole: targetUser?.role || null,
        content: sanitizedContent,
        fileUrl: fileUrl || null,
        fileName: fileName || null,
        fileType: fileType || null,
        fileSize: fileSize ? Number(fileSize) : null,
        replyTo: replyTo && replyTo.content ? {
          id: replyTo.id,
          senderName: replyTo.senderName,
          content: replyTo.content,
        } : null,
        read: false,
      },
    });

    // Real-Time Socket.io Dispatch
    try {
      if (type === 'ANNOUNCEMENT') {
        io.emit('new_announcement', message);
      } else if (type === 'COMMUNITY') {
        io.emit('new_community_message', message);
        if (recipientId) {
          io.to(`user_${recipientId}`).emit('new_community_message', message);
        }
        io.to('role_ADMIN').emit('new_community_message', message);
        io.to('role_MODERATOR').emit('new_community_message', message);
      } else if (type === 'DIRECT') {
        // Send to recipient's private user room
        io.to(`user_${recipientId}`).emit('new_direct_message', message);
        // Send to sender's private user room as well (for multi-tab / other active devices)
        io.to(`user_${user.id}`).emit('new_direct_message', message);
        // Send to Admin audit room for real-time compliance oversight
        io.to('role_ADMIN').emit('admin_message_audit', message);
      }
    } catch (socketErr) {
      logger.warn(`Socket dispatch warning: ${socketErr.message}`);
    }

    return res.status(201).json({
      success: true,
      message,
    });
  } catch (error) {
    logger.error(`Post Chat Message Error: ${error.message}`);
    return res.status(500).json({ success: false, error: 'SEND_FAILED', message: 'Failed to send chat message.' });
  }
}

/**
 * GET /api/v1/chat/community/threads
 * Returns list of user support threads for Staff (Admin/Moderator)
 */
export async function getCommunityThreads(req, res) {
  try {
    const userRole = req.user.role;
    if (userRole !== 'ADMIN' && userRole !== 'MODERATOR') {
      return res.status(403).json({ success: false, error: 'FORBIDDEN', message: 'Staff access required.' });
    }

    const messages = await prisma.chatMessage.findMany({
      where: { type: 'COMMUNITY' },
      orderBy: { createdAt: 'desc' },
      take: 1000,
    });

    const threadMap = new Map();

    for (const msg of messages) {
      let threadUserId = null;
      let threadUserName = null;
      let threadUserRole = null;

      if (msg.senderRole !== 'ADMIN' && msg.senderRole !== 'MODERATOR') {
        threadUserId = msg.senderId;
        threadUserName = msg.senderName;
        threadUserRole = msg.senderRole;
      } else if (msg.recipientId) {
        threadUserId = msg.recipientId;
        threadUserName = msg.recipientName || 'User';
        threadUserRole = msg.recipientRole || 'USER';
      }

      if (!threadUserId) continue;

      if (!threadMap.has(threadUserId)) {
        threadMap.set(threadUserId, {
          id: threadUserId,
          name: threadUserName,
          role: threadUserRole,
          lastMessage: msg.content || msg.fileName || 'Attachment',
          lastSenderName: msg.senderName,
          lastMessageAt: msg.createdAt,
          unread: !msg.read && msg.senderRole !== 'ADMIN' && msg.senderRole !== 'MODERATOR',
        });
      } else {
        const th = threadMap.get(threadUserId);
        if (!th.unread && !msg.read && msg.senderRole !== 'ADMIN' && msg.senderRole !== 'MODERATOR') {
          th.unread = true;
        }
      }
    }

    const userIds = Array.from(threadMap.keys());
    if (userIds.length > 0) {
      const users = await prisma.user.findMany({
        where: { id: { in: userIds } },
        select: { id: true, name: true, role: true, avatar: true, department: true },
      });
      const uMap = new Map(users.map((u) => [u.id, u]));
      for (const [id, th] of threadMap.entries()) {
        const u = uMap.get(id);
        if (u) {
          th.name = u.name || th.name;
          th.role = u.role || th.role;
          th.avatar = u.avatar || null;
          th.department = u.department || null;
        }
      }
    }

    return res.status(200).json({
      success: true,
      threads: Array.from(threadMap.values()),
    });
  } catch (error) {
    logger.error(`Get Community Threads Error: ${error.message}`);
    return res.status(500).json({ success: false, error: 'THREADS_FAILED', message: 'Failed to retrieve community threads.' });
  }
}

/**
 * GET /api/v1/chat/contacts
 * Returns safe list of active platform members for 1-on-1 direct messaging (name, role, avatar only)
 */
export async function getContacts(req, res) {
  try {
    const user = req.user;
    
    // Allow all active platform members (excluding current user)
    const contactWhere = {
      id: { not: user.id },
      status: 'ACTIVE',
    };

    const contacts = await prisma.user.findMany({
      where: contactWhere,
      select: { id: true, name: true, role: true, department: true, avatar: true },
      orderBy: [{ role: 'asc' }, { name: 'asc' }],
      take: 100,
    });

    // Query unread direct messages sent to current user
    const unreadMessages = await prisma.chatMessage.findMany({
      where: {
        type: 'DIRECT',
        recipientId: user.id,
        read: false,
      },
      select: { senderId: true },
    });

    const unreadCountBySender = {};
    for (const m of unreadMessages) {
      unreadCountBySender[m.senderId] = (unreadCountBySender[m.senderId] || 0) + 1;
    }

    const contactsWithUnread = contacts.map((c) => ({
      ...c,
      unreadCount: unreadCountBySender[c.id] || 0,
    }));

    return res.status(200).json({
      success: true,
      contacts: contactsWithUnread,
    });
  } catch (error) {
    logger.error(`Get Chat Contacts Error: ${error.message}`);
    return res.status(500).json({ success: false, error: 'CONTACTS_FAILED', message: 'Failed to retrieve contacts list.' });
  }
}

/**
 * GET /api/v1/chat/unread-count
 * Returns badge count for unread direct messages, announcements, and support desk tickets
 */
export async function getUnreadCount(req, res) {
  try {
    const userId = req.user.id;
    const userRole = req.user.role;
    const isStaff = userRole === 'ADMIN' || userRole === 'MODERATOR';

    // 1. Unread direct messages
    const directCount = await prisma.chatMessage.count({
      where: {
        type: 'DIRECT',
        recipientId: userId,
        read: false,
      },
    });

    // 2. Unread announcements
    const lastViewed = await getLastAnnouncementRead(userId);
    let announcementCount = 0;
    if (lastViewed) {
      announcementCount = await prisma.chatMessage.count({
        where: {
          type: 'ANNOUNCEMENT',
          createdAt: { gt: lastViewed },
        },
      });
    } else {
      const threshold = req.user.createdAt ? new Date(req.user.createdAt) : new Date(Date.now() - 7 * 24 * 60 * 60 * 1000);
      announcementCount = await prisma.chatMessage.count({
        where: {
          type: 'ANNOUNCEMENT',
          createdAt: { gt: threshold },
        },
      });
    }

    // 3. Unread community support messages (seen only when viewed)
    let communityCount = 0;
    if (isStaff) {
      communityCount = await prisma.chatMessage.count({
        where: {
          type: 'COMMUNITY',
          senderRole: { notIn: ['ADMIN', 'MODERATOR'] },
          read: false,
        },
      });
    } else {
      communityCount = await prisma.chatMessage.count({
        where: {
          type: 'COMMUNITY',
          senderRole: { in: ['ADMIN', 'MODERATOR'] },
          OR: [
            { recipientId: userId },
            { recipientId: null },
          ],
          read: false,
        },
      });
    }

    const count = directCount + announcementCount + communityCount;

    return res.status(200).json({
      success: true,
      count,
      directCount,
      announcementCount,
      communityCount,
    });
  } catch (error) {
    return res.status(200).json({ success: true, count: 0, directCount: 0, announcementCount: 0, communityCount: 0 });
  }
}

/**
 * POST /api/v1/chat/read-announcements
 * Marks all announcements as read for the current user
 */
export async function markAnnouncementsRead(req, res) {
  try {
    const userId = req.user.id;
    await setLastAnnouncementRead(userId);
    return res.status(200).json({ success: true, message: 'Announcements marked as read' });
  } catch (error) {
    logger.error(`Mark Announcements Read Error: ${error.message}`);
    return res.status(500).json({ success: false, error: 'UPDATE_FAILED', message: 'Failed to mark announcements as read.' });
  }
}

/**
 * GET /api/v1/chat/notifications
 * Returns list of recent messages across Announcements, Support Desk, and Direct Messages for the Mailbox dropdown
 */
export async function getNotifications(req, res) {
  try {
    const userId = req.user.id;
    const userRole = req.user.role;
    const isStaff = userRole === 'ADMIN' || userRole === 'MODERATOR';
    const lastViewed = (await getLastAnnouncementRead(userId)) || (req.user.createdAt ? new Date(req.user.createdAt) : new Date(Date.now() - 7 * 24 * 60 * 60 * 1000));

    // 1. Recent global announcements
    const recentAnnouncements = await prisma.chatMessage.findMany({
      where: { type: 'ANNOUNCEMENT' },
      orderBy: { createdAt: 'desc' },
      take: 4,
    });

    // 2. Recent community support desk messages
    let commWhere = { type: 'COMMUNITY' };
    if (!isStaff) {
      commWhere = {
        type: 'COMMUNITY',
        OR: [{ recipientId: userId }, { senderId: userId }],
      };
    }
    const recentCommunity = await prisma.chatMessage.findMany({
      where: commWhere,
      orderBy: { createdAt: 'desc' },
      take: 4,
    });

    // 3. Recent 1-on-1 direct messages
    const recentDMs = await prisma.chatMessage.findMany({
      where: {
        type: 'DIRECT',
        OR: [{ recipientId: userId }, { senderId: userId }],
      },
      orderBy: { createdAt: 'desc' },
      take: 4,
    });

    const items = [
      ...recentAnnouncements.map((a) => ({
        id: a.id,
        senderName: a.senderName,
        senderRole: a.senderRole,
        title: `Announcement: ${a.senderName}`,
        message: a.content || a.fileName || 'Attachment',
        channelType: 'ANNOUNCEMENT',
        link: '/inbox?tab=ANNOUNCEMENT',
        read: new Date(a.createdAt) <= new Date(lastViewed),
        createdAt: a.createdAt,
      })),
      ...recentCommunity.map((c) => ({
        id: c.id,
        senderName: c.senderName,
        senderRole: c.senderRole,
        title: `Support Desk: ${c.senderName}`,
        message: c.content || c.fileName || 'Attachment',
        channelType: 'COMMUNITY',
        link: '/inbox?tab=COMMUNITY',
        read: c.senderId === userId ? true : c.read,
        createdAt: c.createdAt,
      })),
      ...recentDMs.map((dm) => ({
        id: dm.id,
        senderName: dm.senderName,
        senderRole: dm.senderRole,
        title: `Direct: ${dm.senderName}`,
        message: dm.content || dm.fileName || 'Attachment',
        channelType: 'DIRECT',
        link: '/inbox?tab=DIRECT',
        read: dm.senderId === userId ? true : dm.read,
        createdAt: dm.createdAt,
      })),
    ].sort((a, b) => new Date(b.createdAt) - new Date(a.createdAt)).slice(0, 10);

    const unreadCount = items.filter((item) => !item.read).length;

    return res.status(200).json({
      success: true,
      messages: items,
      notifications: items,
      unreadCount,
    });
  } catch (error) {
    logger.error(`Get Inbox Recent Messages Error: ${error.message}`);
    return res.status(200).json({ success: true, messages: [], notifications: [], unreadCount: 0 });
  }
}